> fix key regex validation issues by replacing it with actual testing of the key.
> formatting
