+++ backend/accounts/validation.py
@@ -8,28 +8,16 @@ without the cryptography dependency in import path.
 import re
 from typing import Tuple
 
-# Compiled regex patterns for provider-specific key format validation.
-# OpenAI: "sk-" followed by 20+ alphanumeric, underscore, or hyphen chars.
-#   Covers legacy keys (sk-...), project keys (sk-proj-...), service account keys
-#   (sk-svcacct-...), and admin keys (sk-admin-...).
-# Gemini keys: start with "AIza" followed by 10+ alphanumeric/dash/underscore chars.
-OPENAI_KEY_PATTERN = re.compile(r"^sk-[A-Za-z0-9_-]{20,}$")
-GEMINI_KEY_PATTERN = re.compile(r"^AIza[A-Za-z0-9_-]{10,}$")
+# ── Provider-agnostic format validation ──────────────────────────
 
 
 def validate_key_format(provider: str, key: str) -> Tuple[bool, str]:
     """
-    Two-tier key validation.
-
-    Tier 1 — Syntactic (instant, no network):
-    - Check prefix matches known patterns via regex
-    - Check minimum length
-    - Check no whitespace
-
-    Tier 2 — Semantic (network call):
-    - Handled by test_provider_connection() in views.py
-
-    Returns (is_valid, error_message).
+    Simple format validation: non-empty, no whitespace, minimum length.
+    Provider-specific regex patterns are intentionally NOT used — they
+    break when providers change their key format (e.g. new Gemini keys
+    start with 'AQ.' instead of 'AIza').  The real validation happens
+    via test_provider_connection() which actually calls the API.
     """
     if not key or not key.strip():
         return False, "API key is required"
@@ -37,22 +25,104 @@ def validate_key_format(provider: str, key: str) -> Tuple[bool, str]:
     if key != key.strip():
         return False, "API key contains leading or trailing whitespace"
 
-    if provider in ("openai", "morpheus"):
-        if not OPENAI_KEY_PATTERN.match(key):
-            return False, (
-                f"{provider.title()} keys typically start with 'sk-' "
-                "followed by at least 20 chars (letters, digits, underscores, hyphens)"
-            )
-
-    elif provider == "gemini":
-        if not GEMINI_KEY_PATTERN.match(key):
-            return False, (
-                "Gemini keys typically start with 'AIza' "
-                "followed by at least 10 characters"
-            )
-
-    elif provider == "custom":
-        if len(key) < 8:
-            return False, "Custom API key must be at least 8 characters"
+    if len(key) < 8:
+        return False, "API key must be at least 8 characters"
 
     return True, ""
+
+
+# ── API key sanitization ─────────────────────────────────────────
+
+AUTH_FAILURE_PATTERNS = [
+    r"\b401\b",
+    r"\b403\b",
+    "http 401",
+    "http 403",
+    "status 401",
+    "status 403",
+    "401 unauthorized",
+    "403 forbidden",
+    "unauthorized",
+    "forbidden",
+    "invalid api key",
+    "authentication failed",
+    "permission denied",
+    "not authorized",
+    "auth error",
+    "invalid authentication credentials",
+    "api key not found",
+    "credential",
+]
+_AUTH_FAILURE_REGEX = [re.compile(p, re.IGNORECASE) for p in AUTH_FAILURE_PATTERNS]
+
+
+def should_auto_disable(error_message: str) -> bool:
+    return any(patt.search(error_message) for patt in _AUTH_FAILURE_REGEX)
+
+
+def sanitize_error_message(error_msg: str, api_key: str) -> str:
+    if not api_key or not error_msg:
+        return error_msg
+    pattern = re.escape(api_key)
+    error_msg = re.sub(pattern, "***", error_msg, flags=re.IGNORECASE)
+    if len(api_key) >= 12:
+        prefix = re.escape(api_key[:8])
+        suffix = re.escape(api_key[-4:])
+        truncated_pattern = prefix + r".*?" + suffix
+        error_msg = re.sub(truncated_pattern, "***", error_msg, flags=re.IGNORECASE)
+    return error_msg
+
+
+# ── Provider connection testing ──────────────────────────────────
+
+
+def test_provider_connection(provider: str, api_key: str, base_url: str = "") -> tuple:
+    """
+    Test an API key by making a minimal request to the provider.
+    Returns (True, message) on success, (False, error_message) on failure.
+    Error messages have the API key stripped to avoid leaking secrets.
+    """
+    try:
+        if provider in ("openai", "custom", "morpheus"):
+            from openai import OpenAI
+
+            client = OpenAI(
+                api_key=api_key,
+                base_url=base_url or None,
+                timeout=15,
+            )
+            models = client.models.list()
+            first_model = next((m.id for m in models if m.id), "gpt-4o-mini")
+            client.chat.completions.create(
+                model=first_model,
+                messages=[{"role": "user", "content": "test"}],
+                max_tokens=1,
+            )
+            return True, "Connected successfully."
+
+        elif provider == "gemini":
+            from google import genai
+
+            gemini_client = genai.Client(api_key=api_key)
+            gemini_models = list(gemini_client.models.list())
+            first_model = next(
+                (
+                    m.name.split("/")[-1]
+                    for m in gemini_models
+                    if m.name and m.name.startswith("models/gemini")
+                ),
+                "gemini-2.5-flash",
+            )
+            gemini_client.models.generate_content(
+                model=first_model,
+                contents="test",
+                config={"max_output_tokens": 1},
+            )
+            return True, "Connected to Gemini API."
+
+        else:
+            return False, f"Unknown provider: {provider}"
+
+    except Exception as e:
+        error_msg = sanitize_error_message(str(e), api_key)
+        return False, error_msg
